requirements.txt
numpy>=1.21.0
scipy>=1.7.0
Pillow>=9.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
jinja2>=3.1.2
pandas>=1.5.0
tqdm
requests
pandas
flask